Improves quality of unit and integration tests
- reorgs the integration_tests package into /tests
- renames the tests to test_ to be picked up via discovery
- extracts RbacClient into its own class for reuse
- replaces specific test file arg with discover to load all tests in dir
- extracts container startup logic, key gen into a test helper class
- renames former TestBlockchain object to TestOrgHierarchy to allow for
  creation of additional blockchain-oriented integration test files
- removes proxy args from Dockerfile for cleanup and readability
