+++ addressing/rbac_addressing/__tests__/test_addresser.py
@@ -85,7 +85,6 @@ class TestAddresser(unittest.TestCase):
             "The SysAdmin Admins address created must be found to "
             "be a SysAdmin Admins address.")
 
-
     def test_role_addresses(self):
         """Tests the Role address creation functions as well as the
         address_is function.
@@ -165,7 +164,6 @@ class TestAddresser(unittest.TestCase):
             "The Role Tasks address created must be "
             "found to be a Role Tasks address.")
 
-
     def test_task_addresses(self):
         """Tests the Task address creation functions as well as the
         address_is function.
@@ -216,7 +214,6 @@ class TestAddresser(unittest.TestCase):
             "The Task Admins address created must be "
             "found to be a Task Admins address.")
 
-
     def test_proposal_addresses(self):
         """Tests the Proposal address creation function as well as the
         address_is function.
@@ -239,7 +236,6 @@ class TestAddresser(unittest.TestCase):
             AddressSpace.PROPOSALS,
             "The Proposals address created must be found "
             "to be a Proposals address.")
-
 
     def test_users_address(self):
         """Tests the Users address creation function as well as the
