+++ docker-compose.yaml
@@ -31,10 +31,6 @@ services:
     build:
       context: .
       dockerfile: ./server/Dockerfile
-      args:
-        - http_proxy
-        - https_proxy
-        - no_proxy
     container_name: rbac-server
     image: rbac-server-production:${ISOLATION_ID}
     ports:
@@ -57,10 +53,6 @@ services:
     build:
       context: .
       dockerfile: ./processor/Dockerfile
-      args:
-        - http_proxy
-        - https_proxy
-        - no_proxy
     container_name: rbac-tp
     image: rbac-tp-production:${ISOLATION_ID}
     environment:
@@ -71,10 +63,6 @@ services:
     build:
       context: .
       dockerfile: ./ledger_sync/Dockerfile
-      args:
-        - http_proxy
-        - https_proxy
-        - no_proxy
     container_name: rbac-ledger-sync
     image: rbac-ledger-sync-production:${ISOLATION_ID}
     depends_on:
@@ -92,10 +80,6 @@ services:
     build:
       context: .
       dockerfile: ./ui/Dockerfile
-      args:
-        - http_proxy
-        - https_proxy
-        - no_proxy
     container_name: rbac-ui
     image: rbac-ui-production:${ISOLATION_ID}
     ports:
